TEA-61. API для лидерборда
